+++ README.md
@@ -108,8 +108,8 @@ Your finished project must include all of the following requirements:
 Display a list of the plants from the server. This should be done in the class component `PlantList`.
 
 - [ ] In the `PlantList` class component, fetch data from the server you now have running - the data can be fetched from `http://localhost:3333/plants`
-- [ ] Set the data to a state property called `this.state.plants`
-- [ ] The render function is already built and styled. Once the data is on the state, you will see the list of plants, and you will have the functionality to add a plant to the cart
+- [ x] Set the data to a state property called `this.state.plants`
+- [ x] The render function is already built and styled. Once the data is on the state, you will see the list of plants, and you will have the functionality to add a plant to the cart
 
 #### Shopping Cart
