+++ client/src/components/CheckoutForm.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import {useForm} from '../hooks/useForm'
 
 const initialValue = {
   firstName: "",
@@ -14,17 +15,8 @@ const initialValue = {
 // and replace the necessary stateful logic from CheckoutForm with the hook
 
 const CheckoutForm = (props) => {
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-  const [values, setValues] = useState(initialValue);
 
-  const handleChanges = (e) => {
-    setValues({ ...values, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowSuccessMessage(true);
-  };
+  const [values, setValues, handleChanges, handleSubmit, showSuccessMessage ] = useForm(initialValue);
 
   return (
     <>
@@ -68,7 +60,8 @@ const CheckoutForm = (props) => {
         </label>
         <button>Checkout</button>
       </form>
-
+      {console.log(values)}
+      {console.log(showSuccessMessage)}
       {showSuccessMessage && (
         <div className="success-message" data-testid="successMessage">
           <p>
