+++ client/src/components/PlantList.js
@@ -2,6 +2,19 @@ import React, { Component } from "react";
 import axios from "axios";
 
 export default class PlantList extends Component {
+  constructor() {
+    super();
+    this.state = {
+      plants: [],
+    };
+  }
+
+  componentDidMount() {
+    axios
+      .get("http://localhost:3333/plants")
+      .then((res) => this.setState({plants: res.data.plantsData}))
+      .catch(err => console.log(err));
+  }
   // add state with a property called "plants" - initialize as an empty array
 
   // when the component mounts:
@@ -11,6 +24,7 @@ export default class PlantList extends Component {
   /*********  DON'T CHANGE ANYTHING IN THE RENDER FUNCTION *********/
   render() {
     return (
+      
       <main className="plant-list">
         {this.state?.plants?.map((plant) => (
           <div className="plant-card" key={plant.id}>
